Add bundled add16 and j16 instructions
I still haven't tested the semantics of the bundled/parallel
instructions yet, so I'm not sure if this is entirely correct, but at
least now it's possible to add these bundled instructions without having
to duplicate all the existing 16-bit instructions.

To avoid accidentally decoding a 32-bit instruction after a 16-bit
bundle prefix instruction, require the "bundle=0" constraint to be added
to all 32-bit instructions.