Add partial fix for delay slot issues
This isn't a full fix because the delayslot directive only lets you set
the minimum number of bytes in the delay slot--not the number of
instructions. So for cases where an x.d2 instruction is followed by a
4-byte instruction or instruction bundle, only one instruction/bundle
will be placed in the delay slot, which is not how the hardware works.